add modificarComercial & modificarCliente

+++ zzz_examen_mst/src/main/java/unir/exa/ventas/modelo/entity/Comercial.java
@@ -21,6 +21,6 @@ public class Comercial {
 	private String nombre;
 	private String apellido1;
 	private String apellido2;
-	private int comision;
+	private double comision;
 
 }
